Symmetric boundary conditions now work for the distributed case.
Introduced a "code transformation" step that can transform AcTaskDefinitions to a different set of AcTaskDefinitions.
This is used to transform a HaloExchange(boundconds) -> {HaloExchange, Boundconds(boundconds)}

Boundary conditions now receive the correct boundary at which they are in the distributed case and their input region reflects this.
If the boundary condition depends on a halo exchange, this is handled by AcGridBuildTaskGraph just like any other data dependency.
(The output region of the halo exchange will overlap the input region of the boundary condition operation).

Boundary condition tasks are (at the moment) defined for all boundaries, support for separate boundconds for separate boundaries is still missing.