FIX: uninitialized memory in AcMeshes were causing NaNs to propagate all the way to the integration kernels
While distributing an initialized mesh, the receiving processes allocated a host mesh without initializing it, then fill it with the received values.
The issue is that the only values that are propagated are the domain values, which means the ghost zone remains uninitialized.

This commit initializes the memory in the host mesh. I've also initialized some other memory that was being allocated.
It may also be worth considering if acGridLoadMesh should also populate the ghost zone.